requirements.txt
requests>=2.31.0
ijson>=3.2.0
pyarrow[parquet]>=12.0.0
s3fs>=2024.4.0
boto3>=1.29.0
PyYAML>=6.0
backoff>=2.2.1
structlog>=23.1.0
pytest>=7.4.0
tenacity>=8.2.0
python-dotenv>=1.0.0
psutil>=5.9.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
browser-cookie3>=0.19.1 